+++ zk_shell/keys.py
@@ -1,6 +1,47 @@
 """ helpers for JSON keys DSL """
 
+import copy
 import re
+
+
+def container_for_key(key):
+    """ Determines what type of container is needed for `key` """
+    try:
+        int(key)
+        return []
+    except ValueError:
+        return {}
+
+
+def safe_list_set(plist, idx, fill_with, value):
+    """
+    Sets:
+
+    ```
+    plist[idx] = value
+    ```
+
+    If len(plist) is smaller than what idx is trying
+    to dereferece, we first grow plist to get the needed
+    capacity and fill the new elements with fill_with
+    (or fill_with(), if it's a callable).
+    """
+
+    try:
+        plist[idx] = value
+        return
+    except IndexError:
+        pass
+
+    # Fill in the missing positions. Handle negative indexes.
+    end = idx + 1 if idx >= 0 else abs(idx)
+    for _ in range(len(plist), end):
+        if callable(fill_with):
+            plist.append(fill_with())
+        else:
+            plist.append(fill_with)
+
+    plist[idx] = value
 
 
 class Keys(object):
@@ -10,8 +51,11 @@ class Keys(object):
     in template strings
     """
 
-    class Bad(Exception): pass
-    class Missing(Exception): pass
+    class Bad(Exception):
+        pass
+
+    class Missing(Exception):
+        pass
 
     @staticmethod
     def extract(keystr):
@@ -23,6 +67,8 @@ class Keys(object):
         """ validates one key string """
         if re.match(r"\w+(?:\.\w+)*$", keystr) is None:
             raise cls.Bad("Bad key syntax for: %s. Should be: key1.key2..." % (keystr))
+
+        return True
 
     @classmethod
     def from_template(cls, template):
@@ -47,7 +93,6 @@ class Keys(object):
             # plain keys str
             cls.validate_one(keystr)
 
-
     @classmethod
     def fetch(cls, obj, keys):
         """
@@ -68,7 +113,6 @@ class Keys(object):
 
         return current
 
-
     @classmethod
     def value(cls, obj, keystr):
         """
@@ -88,3 +132,66 @@ class Keys(object):
             value = cls.fetch(obj, keystr)
 
         return value
+
+    @classmethod
+    def set(cls, obj, keys, value, fill_list_value=None):
+        """
+        sets the value for the given keys on obj. if any of the given
+        keys does not exist, create the intermediate containers.
+        """
+        current = obj
+        keys_list = keys.split(".")
+
+        for idx, key in enumerate(keys_list, 1):
+            if type(current) == list:
+                # Validate this key works with a list.
+                try:
+                    key = int(key)
+                except ValueError:
+                    raise cls.Missing(key)
+
+            try:
+                # This is the last key, so set the value.
+                if idx == len(keys_list):
+                    if type(current) == list:
+                        safe_list_set(
+                            current,
+                            key,
+                            lambda: copy.copy(fill_list_value),
+                            value
+                        )
+                    else:
+                        current[key] = value
+
+                    # done.
+                    return
+
+                # More keys left, ensure we have a container for this key.
+                if type(key) == int:
+                    try:
+                        current[key]
+                    except IndexError:
+                        # Create a list for this key.
+                        cnext = container_for_key(keys_list[idx])
+                        if type(cnext) == list:
+                            def fill_with():
+                                return []
+                        else:
+                            def fill_with():
+                                return {}
+
+                        safe_list_set(
+                            current,
+                            key,
+                            fill_with,
+                            [] if type(cnext) == list else {}
+                        )
+                else:
+                    if key not in current:
+                        # Create a list for this key.
+                        current[key] = container_for_key(keys_list[idx])
+
+                # Move on to the next key.
+                current = current[key]
+            except (IndexError, KeyError, TypeError):
+                raise cls.Missing(key)