+++ zk_shell/tests/test_json_cmds.py
@@ -5,7 +5,7 @@
 from collections import defaultdict
 import json
 
-from .shell_test_case import  ShellTestCase
+from .shell_test_case import ShellTestCase
 
 
 # pylint: disable=R0904
@@ -129,3 +129,93 @@ class JsonCmdsTestCase(ShellTestCase):
 
         expected_output = u"%s/b\n%s/c\n" % (self.tests_path, self.tests_path)
         self.assertEqual(expected_output, self.output.getvalue())
+
+    def test_json_set_str(self):
+        """ test setting an str """
+        jsonstr = '{"a": {"b": {"c": {"d": "v1"}}}}'
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        self.shell.onecmd("json_set %s/json a.b.c.d v2 str" % (self.tests_path))
+        self.shell.onecmd("json_get %s/json a.b.c.d" % (self.tests_path))
+
+        self.assertEqual("v2\n", self.output.getvalue())
+
+    def test_json_set_int(self):
+        """ test setting an int """
+        jsonstr = '{"a": {"b": {"c": {"d": "v1"}}}}'
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        self.shell.onecmd("json_set %s/json a.b.c.d 2 int" % (self.tests_path))
+        self.shell.onecmd("json_cat %s/json a.b.c.d" % (self.tests_path))
+
+        expected = {u'a': {u'b': {u'c': {u'd': 2}}}}
+        self.assertEqual(expected, json.loads(self.output.getvalue()))
+
+    def test_json_set_bool(self):
+        """ test setting a bool """
+        jsonstr = '{"a": {"b": {"c": {"d": false}}}}'
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        self.shell.onecmd("json_set %s/json a.b.c.d true bool" % (self.tests_path))
+        self.shell.onecmd("json_cat %s/json a.b.c.d" % (self.tests_path))
+
+        expected = {u'a': {u'b': {u'c': {u'd': True}}}}
+        self.assertEqual(expected, json.loads(self.output.getvalue()))
+
+    def test_json_set_json(self):
+        """ test setting serialized json """
+        jsonstr = '{"a": {"b": {"c": {"d": false}}}}'
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        jstr = json.dumps({'c2': {'d2': True}})
+        self.shell.onecmd("json_set %s/json a.b '%s' json" % (self.tests_path, jstr))
+        self.shell.onecmd("json_cat %s/json" % (self.tests_path))
+
+        expected = {u'a': {u'b': {u'c2': {u'd2': True}}}}
+        self.assertEqual(expected, json.loads(self.output.getvalue()))
+
+    def test_json_set_missing_key(self):
+        """ test setting when an intermediate key is missing """
+        jsonstr = '{"a": {"b": {"c": {"d": "v1"}}}}'
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        self.shell.onecmd("json_set %s/json a.b.c.e v2 str" % (self.tests_path))
+        self.shell.onecmd("json_get %s/json a.b.c.d" % (self.tests_path))
+        self.shell.onecmd("json_get %s/json a.b.c.e" % (self.tests_path))
+
+        self.assertEqual("v1\nv2\n", self.output.getvalue())
+
+    def test_json_set_missing_key_with_list(self):
+        """ test setting when an intermediate key is missing and a list has to be created """
+        jsonstr = '{"a": {}}'
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        self.shell.onecmd("json_set %s/json a.b.3.c.e v2 str" % (self.tests_path))
+        self.shell.onecmd("json_cat %s/json" % (self.tests_path))
+
+        expected = {u'a': {u'b': [{}, {}, {}, {u'c': {u'e': u'v2'}}]}}
+        self.assertEqual(expected, json.loads(self.output.getvalue()))
+
+    def test_json_update_list(self):
+        """ test updating an existing inner list """
+        jsonstr = '{"a": [{}, {"b": 2}, {}]}'
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        self.shell.onecmd("json_set %s/json a.1.b 3 str" % (self.tests_path))
+        self.shell.onecmd("json_cat %s/json" % (self.tests_path))
+
+        expected = {u'a': [{}, {u'b': u'3'}, {}]}
+        self.assertEqual(expected, json.loads(self.output.getvalue()))
+
+    def test_json_set_missing_container(self):
+        """ test set """
+        jsonstr = '{"a": {"b": 2}}'
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        self.shell.onecmd("json_set %s/json a.b1.c1.e1 v2 str" % (self.tests_path))
+        self.shell.onecmd("json_cat %s/json" % (self.tests_path))
+
+        expected = {u'a': {u'b': 2, u'b1': {u'c1': {u'e1': u'v2'}}}}
+        self.assertEqual(expected, json.loads(self.output.getvalue()))
+
+    def test_json_set_bad_json(self):
+        """ test with malformed json """
+        jsonstr = '{"a": {"b": {"c": {"d": "v1"}}}'  # missing closing }
+        self.shell.onecmd("create %s/json '%s'" % (self.tests_path, jsonstr))
+        self.shell.onecmd("json_set %s/json a.b.c.e v2 str" % (self.tests_path))
+        self.shell.onecmd("json_get %s/json a.b.c.d" % (self.tests_path))
+
+        expected = "Path /tests/json has bad JSON.\nPath /tests/json has bad JSON.\n"
+        self.assertEqual(expected, self.output.getvalue())