+++ CHANGES.rst
@@ -1,6 +1,21 @@
 ChangeLog
 =========
 
+1.1.5 (2018-08-12)
+------------------
+
+Bug Handling
+~~~~~~~~~~~~
+
+- skip failing tests with zk 3.5.4
+- drop support for python 3.3
+
+Features
+~~~~~~~~
+
+- add json_set command
+
+
 1.1.4 (2018-04-04)
 ------------------